+++ cm-backend/src/main/java/com/clothing/management/dto/ProductHistoryDTO.java
@@ -8,16 +8,16 @@ import java.util.List;
 
 public class ProductHistoryDTO {
 
-    private  Long id;
+    private  long id;
     private String description;
-    private Long productId;
-    private int quantity;
+    private long productId;
+    private long quantity;
     private Date date;
     @Nullable
     private User user;
     private String comment;
 
-    public ProductHistoryDTO(Long id, String description, Long productId, int quantity, Date date, User user, String comment) {
+    public ProductHistoryDTO(long id, String description, long productId, long quantity, Date date, User user, String comment) {
         this.id = id;
         this.description = description;
         this.productId = productId;
@@ -35,19 +35,19 @@ public class ProductHistoryDTO {
         this.description = description;
     }
 
-    public Long getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(Long productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(long quantity) {
         this.quantity = quantity;
     }
